core/users/serializers.py
from .models import User
from rest_framework import serializers


class RegisterSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ('email', 'password', 'first_name', 'last_name', 'exam_attempted', 'exam_marks')

    def create(self, validated_data):
        return User.objects.create_user(**validated_data)